client/requirements.txt
requests
opencv-python
Pillow